dodaje pliki output_file.c output_file.h i poprawione pliki arg_parser.c arg_paser.h main.c Makefile

+++ arg_parser.h
@@ -6,7 +6,7 @@
 struct arguments {
     char *input;
     int parts;
-    int margin;
+    double margin;
     char *output;
     char *format;
 };
